moved template response to external class
